requirements.txt
pandas
redis
python-dateutil
Flask
requests